feat: Move TrackingRepository to new folder structure (#18268)
