+++ src/script/page/LeftSidebar/panels/Conversations/Conversations.tsx
@@ -34,13 +34,13 @@ import {IntegrationRepository} from 'Repositories/integration/IntegrationReposit
 import {PreferenceNotificationRepository} from 'Repositories/notification/PreferenceNotificationRepository';
 import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {SearchRepository} from 'Repositories/search/SearchRepository';
+import {EventName} from 'Repositories/tracking/EventName';
 import {Preferences} from 'src/script/page/LeftSidebar/panels/Preferences';
 import {StartUI} from 'src/script/page/LeftSidebar/panels/StartUI';
 import {ANIMATED_PAGE_TRANSITION_DURATION} from 'src/script/page/MainContent';
 import {useAppMainState, ViewType} from 'src/script/page/state';
 import {ContentState, ListState} from 'src/script/page/useAppState';
 import {TeamRepository} from 'src/script/team/TeamRepository';
-import {EventName} from 'src/script/tracking/EventName';
 import {UserRepository} from 'src/script/user/UserRepository';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
