+++ src/script/repositories/calling/CallingRepository.ts
@@ -68,6 +68,10 @@ import {EventSource} from 'Repositories/event/EventSource';
 import type {MediaDevicesHandler} from 'Repositories/media/MediaDevicesHandler';
 import type {MediaStreamHandler} from 'Repositories/media/MediaStreamHandler';
 import {MediaType} from 'Repositories/media/MediaType';
+import {isCountlyEnabledAtCurrentEnvironment} from 'Repositories/tracking/Countly.helpers';
+import {EventName} from 'Repositories/tracking/EventName';
+import * as trackingHelpers from 'Repositories/tracking/Helpers';
+import {Segmentation} from 'Repositories/tracking/Segmentation';
 import {flatten} from 'Util/ArrayUtil';
 import {calculateChildWindowPosition} from 'Util/DOM/caculateChildWindowPosition';
 import {isDetachedCallingFeatureEnabled} from 'Util/isDetachedCallingFeatureEnabled';
@@ -94,10 +98,6 @@ import {APIClient} from '../../service/APIClientSingleton';
 import {Core} from '../../service/CoreSingleton';
 import {TeamState} from '../../team/TeamState';
 import type {ServerTimeHandler} from '../../time/serverTimeHandler';
-import {isCountlyEnabledAtCurrentEnvironment} from '../../tracking/Countly.helpers';
-import {EventName} from '../../tracking/EventName';
-import * as trackingHelpers from '../../tracking/Helpers';
-import {Segmentation} from '../../tracking/Segmentation';
 import type {UserRepository} from '../../user/UserRepository';
 import {Warnings} from '../../view_model/WarningsContainer';
