+++ src/script/repositories/conversation/MessageRepository.ts
@@ -66,6 +66,10 @@ import {NOTIFICATION_HANDLING_STATE} from 'Repositories/event/NotificationHandli
 import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
 import type {EventRecord, ReactionMap} from 'Repositories/storage';
+import {UserType} from 'Repositories/tracking/attribute';
+import {EventName} from 'Repositories/tracking/EventName';
+import * as trackingHelpers from 'Repositories/tracking/Helpers';
+import {Segmentation} from 'Repositories/tracking/Segmentation';
 import {
   cancelSendingLinkPreview,
   clearLinkPreviewSendingState,
@@ -105,10 +109,6 @@ import {StatusType} from '../../message/StatusType';
 import {Core} from '../../service/CoreSingleton';
 import {TeamState} from '../../team/TeamState';
 import {ServerTimeHandler} from '../../time/serverTimeHandler';
-import {UserType} from '../../tracking/attribute';
-import {EventName} from '../../tracking/EventName';
-import * as trackingHelpers from '../../tracking/Helpers';
-import {Segmentation} from '../../tracking/Segmentation';
 import {protoFromType} from '../../user/AvailabilityMapper';
 import {UserRepository} from '../../user/UserRepository';
 import {UserState} from '../../user/UserState';
