+++ test/helper/TestFactory.js
@@ -53,6 +53,7 @@ import {SearchRepository} from 'Repositories/search/SearchRepository';
 import {SelfService} from 'Repositories/self/SelfService';
 import {SelfRepository} from 'Repositories/self/SelfRepository';
 import {StorageRepository, StorageService} from 'Repositories/storage';
+import {EventTrackingRepository} from 'Repositories/tracking/EventTrackingRepository';
 import {User} from 'src/script/entity/User';
 import {Core} from 'src/script/service/CoreSingleton';
 import {createStorageEngine, DatabaseTypes} from 'src/script/service/StoreEngineProvider';
@@ -60,7 +61,6 @@ import {TeamRepository} from 'src/script/team/TeamRepository';
 import {TeamService} from 'src/script/team/TeamService';
 import {TeamState} from 'src/script/team/TeamState';
 import {serverTimeHandler} from 'src/script/time/serverTimeHandler';
-import {EventTrackingRepository} from 'src/script/tracking/EventTrackingRepository';
 import {UserRepository} from 'src/script/user/UserRepository';
 import {UserService} from 'src/script/user/UserService';
 import {UserState} from 'src/script/user/UserState';
