+++ test/unit_tests/tracking/HelpersSpec.js
@@ -20,9 +20,9 @@
 import {CONVERSATION_TYPE} from '@wireapp/api-client/lib/conversation/';
 import {createUuid} from 'Util/uuid';
 
+import {ConversationType as ConversationTypeAttribute} from 'Repositories/tracking/attribute';
+import * as trackingHelpers from 'Repositories/tracking/Helpers';
 import {Conversation} from 'src/script/entity/Conversation';
-import {ConversationType as ConversationTypeAttribute} from 'src/script/tracking/attribute';
-import * as trackingHelpers from 'src/script/tracking/Helpers';
 
 describe('trackingHelpers', () => {
   describe('getConversationType', () => {
